requirements.txt
numpy
websocket_client